requirements.txt
flask
dramatiq[redis]
dramatiq-abort[redis]
requests
pyyaml
tqdm
django-environ
python-slugify
pytest
tox

addict
cython
pandas
Pillow==9.5.0
scipy
scikit-learn
matplotlib
opencv-python
pycocotools
scikit-image
seaborn
svgwrite
svg.path
termcolor
timm
ultralytics
yapf==0.3.1

git+https://github.com/cocodataset/panopticapi.git#egg=panopticapi

kornia==0.7.*
torch==1.12.1+cu116
torchvision==0.13.1+cu116
torchaudio==0.12.1
--extra-index-url https://download.pytorch.org/whl/cu116
